+++ R/update.R
@@ -38,12 +38,15 @@ update_docs <- function(convert_vignettes = TRUE, path = ".") {
   reformat_md(paste0(good_path, "/README.md"))
 
   # Update changelog, CoC, License
-  update_file("NEWS.md", path)
+  update_file("NEWS.md", path, first = TRUE)
   update_file("CODE_OF_CONDUCT.md", path)
   update_file("LICENSE.md", path)
   if (need_to_bump_version(path)) {
     update_version_number(path)
     cli::cli_alert_success("Bumped version in documentation footer.")
+  }
+  if (need_to_bump_altdoc_version(path)) {
+    update_altdoc_version_number(path)
   }
 
   # Update functions reference
@@ -71,7 +74,7 @@ update_docs <- function(convert_vignettes = TRUE, path = ".") {
 #         - if it changed: overwrite it
 #         - if it didn't: info message
 
-update_file <- function(file, path = ".") {
+update_file <- function(file, path = ".", first) {
 
   file_message <- if (file == "NEWS.md") {
     "NEWS / Changelog"
@@ -118,7 +121,7 @@ update_file <- function(file, path = ".") {
   }
 
   fs::file_copy(orig_file, docs_file, overwrite = TRUE)
-  reformat_md(docs_file)
+  reformat_md(docs_file, first)
 }
 
 update_version_number <- function(path) {
@@ -135,6 +138,31 @@ update_version_number <- function(path) {
     if (length(footer) != 1) return(invisible)
     old_footer <- get_footer(path)
     new_footer <- gsub(doc_version(path), pkg_version(path), old_footer)
+    index[footer] <- new_footer
+    writeLines(index, "docs/index.html")
+  } else if (doc_type == "mkdocs") {
+    # TODO ? Or is it linked to the github page ?
+  }
+}
+
+update_altdoc_version_number <- function(path) {
+  doc_type <- doc_type(path)
+  if (doc_type %in% c("docute", "docsify")) {
+    index <- .readlines("docs/index.html")
+    index2 <- gsub("\\t", "", index)
+    index2 <- trimws(index2)
+    if (doc_type == "docsify") {
+      footer <- which(grepl("^var footer =", index2))
+    } else if (doc_type == "docute") {
+      footer <- which(grepl("^footer:", index2))
+    }
+    if (length(footer) != 1) return(invisible)
+    old_footer <- get_footer(path)
+    new_footer <- gsub(
+      altdoc_version_in_footer(path),
+      altdoc_version(),
+      old_footer
+    )
     index[footer] <- new_footer
     writeLines(index, "docs/index.html")
   } else if (doc_type == "mkdocs") {